first try at nll for Poisson binomial data
This script is the beginning of an effort to develop a likelihood
function that is appropriate for a dispersal experiment run by Tad
Dallas while he as at UC Davis in the lab of Alan Hastings around
2017. The basic idea is that beetles can disperse to or from one of
two patches and there are "before dispersal" and "after dispersal"
data. However, the origin of beetles censused after dispersal is
unknown and thus "after dispersal" counts must be treated as the sum
of two binomially distributed random variables, representing dispersal
from each patch. Hence, the use of a Poisson binomial model.